requirements.txt
black
isort
mypy
pylint

matplotlib
plexapi
progressbar2
tautulli
urllib3>=2
